refactor(standard): 重构站点信息获取逻辑
- 将 `SiteInfoService` 重命名为 `PublicLogic` 并实现 `PublicService` 接口
- 更新 `PublicController` 中的依赖注入，使用 `PublicLogic` 替代 `SiteInfoService`
- 移除 `RedisConfig` 中未使用的 `@EnableCaching` 注解
- 添加新的 `PublicService` 接口，定义 `getSiteInfo` 方法
